Extract RPC command handlers from AppDelegate into RPCHandlers/
Move 17 inline RPC handler closures out of AppDelegate.startSocketServer()
into focused files under RPCHandlers/, grouped by domain: sessions,
worktrees, terminals, links, and hook events. Each file exports a factory
function returning [String: CommandRouter.Handler] that AppDelegate merges
at startup.

AppDelegate shrinks from 941 to 427 LOC with no behavior changes.

Closes #138
